Moved configuration copy out to IConfigUpdater service
