--- Neovolve.Configuration.DependencyInjection/ConfigUpdateExtensions.cs
@@ -1,163 +1,0 @@
-﻿namespace Neovolve.Configuration.DependencyInjection
-{
-    using System;
-    using System.Reflection;
-    using Microsoft.Extensions.DependencyInjection;
-    using Microsoft.Extensions.Logging;
-
-    internal static partial class ConfigUpdateExtensions
-    {
-        public static void CopyValues<T>(this IServiceProvider serviceProvider, T? injectedConfig, T updatedConfig)
-        {
-            if (injectedConfig == null)
-            {
-                return;
-            }
-
-            if (updatedConfig == null)
-            {
-                return;
-            }
-
-            var targetType = typeof(T);
-            var properties = targetType.GetBindableProperties(true);
-            var logger = GetLogger<T>(serviceProvider);
-
-            // The IOptionsMonitor<T>.OnChange event gets triggered twice on a change notification for file based configuration
-            // To prevent noise in logging and updates to properties, we want to try to detect when an actual change has been made
-            // We only want to set a property value if there is a change in the value, and we want to only log that changes have
-            // occurred the first time a change is detected
-            var changesFound = false;
-
-            foreach (var property in properties)
-            {
-                var propertyUpdated = UpdateProperty(targetType, property, injectedConfig, updatedConfig, logger);
-
-                if (propertyUpdated)
-                {
-                    changesFound = true;
-                }
-            }
-
-            // We have a change to the configuration
-            if (changesFound)
-            {
-                if (logger != null)
-                {
-                    LogConfigChanged(logger, targetType);
-                }
-            }
-        }
-
-        private static bool BothValuesNull(object? oldValue, object? updatedValue)
-        {
-            if (oldValue != null)
-            {
-                return false;
-            }
-
-            if (updatedValue != null)
-            {
-                return false;
-            }
-
-            return true;
-        }
-
-        private static ILogger? GetLogger<T>(IServiceProvider serviceProvider)
-        {
-            var factory = serviceProvider.GetService<ILoggerFactory>();
-            var logger = factory?.CreateLogger<T>();
-
-            return logger;
-        }
-
-        private static bool IsMatchingValue(object? oldValue, object? updatedValue)
-        {
-            if (oldValue == null)
-            {
-                return false;
-            }
-
-            if (updatedValue == null)
-            {
-                return false;
-            }
-
-            if (oldValue.Equals(updatedValue))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        private static bool IsWritable(PropertyInfo property)
-        {
-            if (property.CanWrite == false)
-            {
-                return false;
-            }
-
-            if (property.SetMethod.IsPublic == false)
-            {
-                return false;
-            }
-
-            return true;
-        }
-
-        private static bool UpdateProperty<T>(Type targetType, PropertyInfo property, T injectedConfig, T updatedConfig,
-            ILogger? logger)
-        {
-            if (IsWritable(property) == false)
-            {
-                if (logger != null)
-                {
-                    LogConfigCopyDenied(logger, targetType, property.Name);
-                }
-
-                return false;
-            }
-
-            try
-            {
-                var oldValue = property.GetValue(injectedConfig);
-                var updatedValue = property.GetValue(updatedConfig, null);
-
-                if (BothValuesNull(oldValue, updatedValue))
-                {
-                    // There is no change
-                    return false;
-                }
-
-                if (IsMatchingValue(oldValue, updatedValue))
-                {
-                    // Both values exist but they are the same
-                    return false;
-                }
-
-                property.SetValue(injectedConfig, updatedValue);
-
-                if (logger != null)
-                {
-                    LogConfigPropertyChanged(logger, targetType, property.Name, oldValue, updatedValue);
-                }
-
-                return true;
-            }
-            catch (Exception ex)
-            {
-                // We have failed to copy the value across to the injected config value
-                // We don't want to fail the application so we cannot allow the exception to throw up the stack
-                // We can however attempt to obtain a logger so that we can report the failure
-                if (logger != null)
-                {
-                    LogConfigCopyFail(logger, ex, targetType, property.Name);
-                }
-
-                return false;
-            }
-        }
-    }
-}
